refactor ui widgets
