+++ ros2_unbag/ui/widgets/export_options.py
@@ -20,342 +20,14 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-import inspect
-import os
-
 from PySide6 import QtCore, QtWidgets
 
 from ros2_unbag.core.processors import Processor
 from ros2_unbag.core.routines import ExportRoutine, ExportMode
 
+from .processor_chain import ProcessorChainWidget
 
-class ProcessorChainWidget(QtWidgets.QWidget):
-    """Widget to configure an ordered chain of processors for a topic."""
-
-    def __init__(self, topic_type, available_processors, parent=None):
-        super().__init__(parent)
-        self.topic_type = topic_type
-        self.available_processors = list(available_processors)
-        self.entries = []
-
-        layout = QtWidgets.QVBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.setSpacing(6)
-
-        self.chain_layout = QtWidgets.QVBoxLayout()
-        self.chain_layout.setSpacing(8)
-        layout.addLayout(self.chain_layout)
-
-        self.empty_hint = QtWidgets.QLabel("No processors configured.")
-        self.empty_hint.setStyleSheet("color: gray; font-style: italic;")
-        layout.addWidget(self.empty_hint)
-
-        add_row = QtWidgets.QHBoxLayout()
-        add_row.addStretch()
-        self.add_button = QtWidgets.QPushButton("Add Processor")
-        self.add_button.setCursor(QtCore.Qt.PointingHandCursor)
-        self.add_button.clicked.connect(self.add_entry)
-        add_row.addWidget(self.add_button)
-        layout.addLayout(add_row)
-
-        self._update_empty_hint()
-
-    def add_entry(self, preset=None):
-        entry = _ProcessorEntry(self)
-        self.entries.append(entry)
-        self.chain_layout.addWidget(entry)
-        self._reindex_entries()
-        if preset:
-            entry.apply_config(preset)
-        else:
-            entry.select_default()
-        self._update_empty_hint()
-        return entry
-
-    def remove_entry(self, entry):
-        if entry in self.entries:
-            self.entries.remove(entry)
-            self.chain_layout.removeWidget(entry)
-            entry.deleteLater()
-            self._reindex_entries()
-            self._update_empty_hint()
-
-    def move_entry(self, entry, delta):
-        if entry not in self.entries:
-            return
-        idx = self.entries.index(entry)
-        new_idx = idx + delta
-        if new_idx < 0 or new_idx >= len(self.entries):
-            return
-        self.entries.pop(idx)
-        self.entries.insert(new_idx, entry)
-        self.chain_layout.removeWidget(entry)
-        self.chain_layout.insertWidget(new_idx, entry)
-        self._reindex_entries()
-
-    def get_chain(self):
-        chain = []
-        for entry in self.entries:
-            config = entry.get_config()
-            if config:
-                chain.append(config)
-        return chain
-
-    def set_chain(self, configs):
-        self.clear()
-        for cfg in configs or []:
-            self.add_entry(cfg)
-        if not configs:
-            self._update_empty_hint()
-
-    def clear(self):
-        for entry in list(self.entries):
-            self.remove_entry(entry)
-
-    def _reindex_entries(self):
-        total = len(self.entries)
-        for idx, entry in enumerate(self.entries, start=1):
-            entry.set_index(idx, total)
-
-    def _update_empty_hint(self):
-        self.empty_hint.setVisible(len(self.entries) == 0)
-
-
-class _ProcessorEntry(QtWidgets.QFrame):
-    """Single processor entry row inside the processor chain widget."""
-
-    def __init__(self, chain_widget):
-        super().__init__()
-        self.chain_widget = chain_widget
-        self.arg_inputs = {}
-
-        self.setFrameShape(QtWidgets.QFrame.StyledPanel)
-        self.setFrameShadow(QtWidgets.QFrame.Raised)
-
-        layout = QtWidgets.QVBoxLayout(self)
-        layout.setContentsMargins(8, 8, 8, 8)
-        layout.setSpacing(4)
-
-        header = QtWidgets.QHBoxLayout()
-        header.setSpacing(6)
-
-        self.index_label = QtWidgets.QLabel("1.")
-        self.index_label.setFixedWidth(24)
-        header.addWidget(self.index_label)
-
-        self.combo = QtWidgets.QComboBox()
-        self.combo.addItems(self.chain_widget.available_processors)
-        self.combo.currentTextChanged.connect(self._on_processor_changed)
-        header.addWidget(self.combo, stretch=1)
-
-        self.up_button = QtWidgets.QToolButton()
-        self.up_button.setIcon(self.style().standardIcon(QtWidgets.QStyle.SP_ArrowUp))
-        self.up_button.setAutoRaise(True)
-        self.up_button.setCursor(QtCore.Qt.PointingHandCursor)
-        self.up_button.clicked.connect(lambda: self.chain_widget.move_entry(self, -1))
-        header.addWidget(self.up_button)
-
-        self.down_button = QtWidgets.QToolButton()
-        self.down_button.setIcon(self.style().standardIcon(QtWidgets.QStyle.SP_ArrowDown))
-        self.down_button.setAutoRaise(True)
-        self.down_button.setCursor(QtCore.Qt.PointingHandCursor)
-        self.down_button.clicked.connect(lambda: self.chain_widget.move_entry(self, 1))
-        header.addWidget(self.down_button)
-
-        self.remove_button = QtWidgets.QToolButton()
-        self.remove_button.setIcon(self.style().standardIcon(QtWidgets.QStyle.SP_DialogCloseButton))
-        self.remove_button.setAutoRaise(True)
-        self.remove_button.setCursor(QtCore.Qt.PointingHandCursor)
-        self.remove_button.clicked.connect(lambda: self.chain_widget.remove_entry(self))
-        header.addWidget(self.remove_button)
-
-        layout.addLayout(header)
-
-        self.args_layout = QtWidgets.QFormLayout()
-        self.args_layout.setContentsMargins(0, 4, 0, 0)
-        self.args_layout.setFieldGrowthPolicy(QtWidgets.QFormLayout.ExpandingFieldsGrow)
-        layout.addLayout(self.args_layout)
-
-        self._on_processor_changed(self.combo.currentText())
-
-    def set_index(self, index, total):
-        self.index_label.setText(f"{index}.")
-        self.up_button.setEnabled(index > 1)
-        self.down_button.setEnabled(index < total)
-
-    def select_default(self):
-        if self.chain_widget.available_processors:
-            self.combo.setCurrentIndex(0)
-
-    def apply_config(self, config):
-        name = config.get("name")
-        if name and name in self.chain_widget.available_processors:
-            self.combo.setCurrentText(name)
-        elif name:
-            # Allow configuring processors that were loaded dynamically after widget creation.
-            self.combo.addItem(name)
-            self.combo.setCurrentText(name)
-
-        args = config.get("args", {}) or {}
-        self._on_processor_changed(self.combo.currentText())
-        for arg_name, value in args.items():
-            if arg_name in self.arg_inputs:
-                self.arg_inputs[arg_name].setText(str(value))
-
-    def get_config(self):
-        name = self.combo.currentText()
-        args = {}
-        for arg_name, edit in self.arg_inputs.items():
-            value = edit.text().strip()
-            if value:
-                args[arg_name] = value
-        return {"name": name, "args": args}
-
-    def _clear_args(self):
-        while self.args_layout.rowCount():
-            self.args_layout.removeRow(0)
-        self.arg_inputs = {}
-
-    def _on_processor_changed(self, processor_name):
-        self._clear_args()
-        args = Processor.get_args(self.chain_widget.topic_type, processor_name)
-        if not args:
-            return
-
-        for arg_name, (param, doc) in args.items():
-            label = QtWidgets.QLabel()
-            label.setText(f"{arg_name} (optional)" if param.default != inspect.Parameter.empty else arg_name)
-
-            parts = []
-            if doc:
-                parts.append(doc)
-            if param.default != inspect.Parameter.empty:
-                parts.append(f"default: {param.default}")
-            if param.annotation != inspect.Parameter.empty:
-                annotation = getattr(param.annotation, "__name__", str(param.annotation))
-                parts.append(f"Type: {annotation}")
-            placeholder_text = " — ".join(parts)
-
-            arg_edit = QtWidgets.QLineEdit()
-            arg_edit.setPlaceholderText(placeholder_text)
-
-            self.args_layout.addRow(label, arg_edit)
-            self.arg_inputs[arg_name] = arg_edit
-
-
-class TopicSelector(QtWidgets.QWidget):
-    # Widget to display and select available topics from the bag
-
-    def __init__(self, bag_reader):
-        """
-        Initialize TopicSelector with a BagReader, retrieve topics and message counts, and build the UI.
-
-        Args:
-            bag_reader: BagReader instance for the ROS2 bag.
-
-        Returns:
-            None
-        """
-        super().__init__()
-        self.bag_reader = bag_reader
-        self.topics = self.bag_reader.get_topics()
-        self.message_counts = self.bag_reader.get_message_count()
-        self.checkboxes = {}
-        self.select_all_state = True
-
-        self.init_ui()
-
-    def init_ui(self):
-        """
-        Build the topic selection UI: group topics by message type with checkboxes and message count labels.
-
-        Args:
-            None
-
-        Returns:
-            None
-        """
-        layout = QtWidgets.QVBoxLayout()
-
-        # Create checkboxes grouped by message type
-        for msg_type, topic_list in sorted(self.topics.items()):
-            group_box = QtWidgets.QGroupBox(msg_type)
-            group_layout = QtWidgets.QVBoxLayout()
-
-            for topic in sorted(topic_list):
-                container = QtWidgets.QWidget()
-                h_layout = QtWidgets.QHBoxLayout()
-                h_layout.setContentsMargins(0, 0, 0, 0)
-
-                checkbox = QtWidgets.QCheckBox()
-                checkbox.setCursor(QtCore.Qt.PointingHandCursor)
-                label = QtWidgets.QLabel(topic)
-                label.setCursor(QtCore.Qt.PointingHandCursor)
-                label.mousePressEvent = self._make_label_toggle_cb(checkbox)
-                count_label = QtWidgets.QLabel(str(self.message_counts.get(topic, 0))+ " Messages")
-                count_label.setAlignment(QtCore.Qt.AlignRight | QtCore.Qt.AlignVCenter)
-
-                h_layout.addWidget(checkbox)
-                h_layout.addWidget(label)
-                h_layout.addStretch()
-                h_layout.addWidget(count_label)
-
-                container.setLayout(h_layout)
-                group_layout.addWidget(container)
-                self.checkboxes[topic] = checkbox
-
-            group_box.setLayout(group_layout)
-            layout.addWidget(group_box)
-
-        # Select All / Deselect All button
-        self.select_all_button = QtWidgets.QPushButton("Select All")
-        self.select_all_button.clicked.connect(self.toggle_select_all)
-        layout.addWidget(self.select_all_button)
-
-        self.setLayout(layout)
-
-    def toggle_select_all(self):
-        """
-        Toggles the selection state of all checkboxes in the widget.
-
-        Args:
-            None
-
-        Returns:
-            None
-        """
-        for cb in self.checkboxes.values():
-            cb.setChecked(self.select_all_state)
-        self.select_all_state = not self.select_all_state
-        self.select_all_button.setText("Deselect All" if not self.select_all_state else "Select All")
-
-    def _make_label_toggle_cb(self, checkbox):
-        """
-        Creates a callback function that toggles the state of the given checkbox.
-
-        Args:
-            checkbox: QCheckBox instance to toggle.
-
-        Returns:
-            function: Callback function for mousePressEvent.
-        """
-        def toggle(_):
-            checkbox.toggle()
-        return toggle
-    
-    def get_selected_topics(self):
-        """
-        Return a list of topics whose checkboxes are currently checked.
-
-        Args:
-            None
-
-        Returns:
-            list: List of selected topic names.
-        """
-        return [
-            topic for topic, cb in self.checkboxes.items() if cb.isChecked()
-        ]
+__all__ = ["ExportOptions"]
 
 
 class ExportOptions(QtWidgets.QWidget):
@@ -864,3 +536,4 @@ class ExportOptions(QtWidgets.QWidget):
             self, "Select Directory")
         if directory:
             edit.setText(directory)
+
